feat: Support restricted keycode
Add support for a keycode "restricted" flag, which is set by product
code (via interfaces in `nx_keycode`) and is reset via a special
keycode available in both small and full keycode protocols.

Other notable updates:
- It is now easier to use this code in C++ projects
- More strict static analysis is now available, including the use
  of multiple compilers and compiler versions
- Fixed unit tests with memory/address sanitizer errors

+++ nexus/test/test_nexus_keycode_pro_misc.c
@@ -81,7 +81,7 @@ void test_mask_below_message_id__id_below_current_window__no_mask_changes(void)
         if (i != 301)
         {
             TEST_ASSERT_EQUAL_UINT(
-                nexus_keycode_pro_get_full_message_id_flag(i), 0);
+                nexus_keycode_pro_get_full_message_id_flag((uint16_t) i), 0);
         }
     }
 }
